# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "EngineerCMS API",
        "description": "ECMS has every tool to get any job done, so codename for the new ECMS APIs.",
        "version": "1.0.0",
        "contact": {
            "email": "504284@qq.com"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/wx/uploadwxavatar": {
            "post": {
                "tags": [
                    "wx"
                ],
                "description": "post user avatar",
                "operationId": "FroalaController.post wx user avatar",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.SUCCESS"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "articl not found"
                    }
                }
            }
        },
        "/wx/uploadwxeditorimg": {
            "post": {
                "tags": [
                    "wx"
                ],
                "description": "post article img by catalogid",
                "operationId": "FroalaController.post wx artile img by catalogId",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.SUCCESS"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "articl not found"
                    }
                }
            }
        },
        "/wx/uploadwximg": {
            "post": {
                "tags": [
                    "wx"
                ],
                "description": "post article img by catalogid",
                "operationId": "FroalaController.post wx artile img by catalogId",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.SUCCESS"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "articl not found"
                    }
                }
            }
        },
        "/wx/uploadwximgs/{id}": {
            "post": {
                "tags": [
                    "wx"
                ],
                "description": "post article img by catalogid",
                "operationId": "FroalaController.post wx artile img by catalogId",
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "The id of project",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.SUCCESS"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "articl not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        ".SUCCESS": {
            "type": "object"
        },
...
    },
...
}